requirements.txt
python-dotenv
python-telegram-bot
python-telegram-bot[job-queue]
tzdata